Don't use the ohai cookbook to install the plugins.
It forces an additional reload of ohai every run, instead of just when
the plugins change.  Now, we never reload (so the plugins won't be
available until the next run).  This should be fine, as some version
of the plugin will be installed during the build of the base AMI.